+++ src/textual/_compositor.py
@@ -598,6 +598,7 @@ class Compositor:
                 if widget.is_container:
                     # Arrange the layout
                     arrange_result = widget._arrange(child_region.size)
+
                     arranged_widgets = arrange_result.widgets
                     widgets.update(arranged_widgets)
 
@@ -615,6 +616,11 @@ class Compositor:
                     # An offset added to all placements
                     placement_offset = container_region.offset
                     placement_scroll_offset = placement_offset - widget.scroll_offset
+
+                    placements = [
+                        placement.process_offset(size.region, placement_scroll_offset)
+                        for placement in placements
+                    ]
 
                     layers_to_index = {
                         layer_name: index
@@ -643,6 +649,7 @@ class Compositor:
                         z,
                         fixed,
                         overlay,
+                        absolute,
                     ) in reversed(placements):
                         layer_index = get_layer_index(sub_widget.layer, 0)
                         # Combine regions with children to calculate the "virtual size"
@@ -656,17 +663,6 @@ class Compositor:
                             )
 
                         widget_order = order + ((layer_index, z, layer_order),)
-
-                        if overlay:
-                            styles = sub_widget.styles
-                            has_rule = styles.has_rule
-                            if has_rule("constrain_x") or has_rule("constrain_y"):
-                                widget_region = widget_region.constrain(
-                                    styles.constrain_x,
-                                    styles.constrain_y,
-                                    styles.margin,
-                                    no_clip,
-                                )
 
                         if widget._cover_widget is None:
                             add_widget(
@@ -712,22 +708,6 @@ class Compositor:
 
             elif visible:
                 # Add the widget to the map
-
-                if widget.absolute_offset is not None:
-                    margin = styles.margin
-                    region = region.at_offset(widget.absolute_offset + margin.top_left)
-                    region = region.translate(
-                        styles.offset.resolve(region.grow(margin).size, size)
-                    )
-                has_rule = styles.has_rule
-                if has_rule("constrain_x") or has_rule("constrain_y"):
-                    region = region.constrain(
-                        styles.constrain_x,
-                        styles.constrain_y,
-                        styles.margin,
-                        size.region,
-                    )
-
                 map[widget._render_widget] = _MapGeometry(
                     region,
                     order,